Generic Video Generation/requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.2.0
audiostretchy==1.3.5
blinker==1.9.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Cython==3.0.11
decorator==4.4.2
dtw-python==1.5.3
edge-tts==6.1.12
filelock==3.16.1
fire==0.7.0
Flask==3.1.0
frozenlist==1.4.1
fsspec==2024.10.0
google-api-core==2.21.0
google-auth==2.35.0
google-cloud-texttospeech==2.20.0
googleapis-common-protos==1.65.0
grpcio==1.67.0
grpcio-status==1.67.0
gTTS==2.5.3
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai-whisper==20240930
Pillow==9.4.0
proglog==0.1.10
propcache==0.2.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyttsx3==2.98
regex==2024.9.11
requests==2.32.3
rsa==4.9
scipy==1.14.1
sympy==1.13.1
termcolor==2.5.0
tiktoken==0.8.0
torch==2.5.0
tqdm==4.66.5
triton==3.1.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
whisper-timestamped==1.15.4
yarl==1.15.4
